requirements.txt
CairoSVG==1.0.22
Django==1.10
Pillow==3.4.2
Pyphen==0.9.4
Unidecode==0.04.19
WeasyPrint==0.32
cairocffi==0.7.2
cffi==1.9.1
cssselect==1.0.0
django-debug-toolbar==1.6
django-multiupload==0.5.2
django-summernote==0.8.6
django-uuslug==1.1.8
et-xmlfile==1.0.1
html5lib==0.999999999
jdcal==1.3
lxml==3.6.4
openpyxl==2.4.0
pycparser==2.17
python-slugify==1.2.1
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.2.2
tinycss==0.4
webencodings==0.5
